+++ Project-for-Medical/src/main/java/com/medicare/ProjectforMedical/Dto/DoctorResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -14,5 +16,5 @@ public class DoctorResponse {
     private String name;
     private String image;
     private String speciality;
-    private int yearsOfExp;
+    private BigDecimal yearsOfExp;
 }
